backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic==2.11.2
pydantic-settings==2.8.1
celery[redis]==5.5.1
redis==5.2.1
ffmpeg-python==0.2.0
scipy==1.15.2
numpy==2.2.4
librosa==0.10.2
python-jose[cryptography]==3.3.0
httpx==0.28.1
aiofiles==24.1.0
python-magic==0.4.27
ruff==0.9.10
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
opencv-python-headless==4.10.0.84
scikit-learn==1.6.1
tqdm==4.67.1
